+++ democrasite-frontend/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Bill/Bill";
+export * from "./BillList/BillList";
+export * from "./SignInButton/SignInButton";
