+++ democrasite-frontend/lib/auto/.openapi-generator-ignore
