+++ democrasite-frontend/lib/auto/.openapi-generator/FILES
