+++ democrasite-frontend/lib/auto/.openapi-generator/VERSION
