+++ democrasite-frontend/lib/auto/apis/AuthpasswordApi.ts
