+++ democrasite-frontend/lib/auto/apis/BillsApi.ts
