+++ democrasite-frontend/lib/auto/index.ts
