+++ democrasite-frontend/lib/auto/models/Bill.ts
