+++ democrasite-frontend/lib/auto/models/PatchedBill.ts
