+++ democrasite-frontend/lib/auto/models/PatchedUserDetails.ts
