+++ democrasite-frontend/lib/auto/models/RestAuthDetail.ts
