+++ democrasite-frontend/lib/auto/models/SocialLogin.ts
