+++ democrasite-frontend/lib/auto/models/UserDetails.ts
